+++ sharding-orchestration/sharding-orchestration-core/src/main/java/io/shardingsphere/orchestration/internal/config/AuthenticationOrchestrationListener.java
@@ -19,7 +19,7 @@ package io.shardingsphere.orchestration.internal.config;
 
 import io.shardingsphere.core.event.ShardingEventBusInstance;
 import io.shardingsphere.orchestration.internal.event.config.AuthenticationChangedEvent;
-import io.shardingsphere.orchestration.internal.listener.ListenerManager;
+import io.shardingsphere.orchestration.internal.listener.OrchestrationListener;
 import io.shardingsphere.orchestration.reg.api.RegistryCenter;
 import io.shardingsphere.orchestration.reg.listener.DataChangedEvent;
 import io.shardingsphere.orchestration.reg.listener.EventListener;
@@ -29,7 +29,7 @@ import io.shardingsphere.orchestration.reg.listener.EventListener;
  *
  * @author panjuan
  */
-public final class AuthenticationListenerManager implements ListenerManager {
+public final class AuthenticationOrchestrationListener implements OrchestrationListener {
     
     private final ConfigurationNode configNode;
     
@@ -37,7 +37,7 @@ public final class AuthenticationListenerManager implements ListenerManager {
     
     private final ConfigurationService configService;
     
-    public AuthenticationListenerManager(final String name, final RegistryCenter regCenter) {
+    public AuthenticationOrchestrationListener(final String name, final RegistryCenter regCenter) {
         configNode = new ConfigurationNode(name);
         this.regCenter = regCenter;
         configService = new ConfigurationService(name, regCenter);
