+++ sharding-orchestration/sharding-orchestration-core/src/main/java/io/shardingsphere/orchestration/internal/config/DataSourceOrchestrationListener.java
@@ -19,7 +19,7 @@ package io.shardingsphere.orchestration.internal.config;
 
 import io.shardingsphere.core.event.ShardingEventBusInstance;
 import io.shardingsphere.orchestration.internal.event.config.DataSourceChangedEvent;
-import io.shardingsphere.orchestration.internal.listener.ListenerManager;
+import io.shardingsphere.orchestration.internal.listener.OrchestrationListener;
 import io.shardingsphere.orchestration.internal.state.datasource.DataSourceService;
 import io.shardingsphere.orchestration.reg.api.RegistryCenter;
 import io.shardingsphere.orchestration.reg.listener.DataChangedEvent;
@@ -30,7 +30,7 @@ import io.shardingsphere.orchestration.reg.listener.EventListener;
  *
  * @author panjuan
  */
-public final class DataSourceListenerManager implements ListenerManager {
+public final class DataSourceOrchestrationListener implements OrchestrationListener {
     
     private final ConfigurationNode configNode;
     
@@ -40,7 +40,7 @@ public final class DataSourceListenerManager implements ListenerManager {
     
     private final DataSourceService dataSourceService;
     
-    public DataSourceListenerManager(final String name, final RegistryCenter regCenter, final String shardingSchemaName) {
+    public DataSourceOrchestrationListener(final String name, final RegistryCenter regCenter, final String shardingSchemaName) {
         configNode = new ConfigurationNode(name);
         this.regCenter = regCenter;
         this.shardingSchemaName = shardingSchemaName;
